Allow recoil values to be saved to 2.d.p

+++ src/inputs/Mouse.cpp
@@ -15,7 +15,7 @@ namespace Inputs
         return (GetAsyncKeyState(VK_LBUTTON) & 0x8000);
     }
 
-    void MoveMouseRaw(int dx, int dy)
+    void MoveMouseRaw(float dx, float dy)
     {
         INPUT input = { 0 };
         input.type = INPUT_MOUSE;
